refactor: 비즈니스 로직이 들어간 validator 제거 및 서비스로 이동 (피드백 반영)
