+++ src/main/java/gift/constant/ErrorMessage.java
@@ -16,6 +16,7 @@ public enum ErrorMessage {
     MISSING_TOKEN("헤더에 토큰이 존재하지 않거나 잘못된 형식입니다."),
 
     /* MemberService */
+    DUPLICATED_EMAIL("중복된 이메일입니다."),
     LOGIN_FAILURE("이메일 또는 비밀번호가 일치하지 않습니다."),
 
     /* WishlistService */
