+++ src/main/java/gift/exception/EmailDuplicateException.java
@@ -1,0 +1,16 @@
+package gift.exception;
+
+import gift.constant.ErrorMessage;
+
+public class EmailDuplicateException extends GiftException {
+
+    public EmailDuplicateException() {
+        super(ErrorMessage.DUPLICATED_EMAIL.getMessage());
+    }
+
+    @Override
+    public int getStatusCode() {
+        return 409;
+    }
+
+}
