+++ src/main/java/gift/request/RegisterRequest.java
@@ -1,7 +1,6 @@
 package gift.request;
 
 import gift.domain.member.Member;
-import gift.validation.member.NotDuplicateEmail;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Pattern;
 import org.hibernate.validator.constraints.Length;
@@ -14,7 +13,6 @@ public class RegisterRequest {
 
     @NotBlank
     @Pattern(regexp = "[a-zA-Z0-9_!#$%&’*+/=?`{|}~^.-]+@[a-zA-Z0-9.-]+$", message = "올바른 이메일 형식을 입력해 주세요.")
-    @NotDuplicateEmail
     private String email;
 
     @NotBlank
